create a genData file
